requirements.txt
pillow
pytesseract
pyperclip